+++ CHANGELOG.md
@@ -1,17 +1,39 @@
 # CHANGELOG
 
-## [1.0.3] - 2025-03-20
+## [2.0.0] - 2025-03-22
+
+### Removed
+
+- Unity.Logging dependency
+
+### Added
+
+- Image commands
+
 ### Changed
+
+- `CommandType` -> `LineType`
+- `RegisterStoryControllerDependencies` has better configuration options now
+- `ContinueMessage.Maximally = true` now only sends the text when encountering a command and buffers it; it will be published on the next continue message
+
+## [1.0.3] - 2025-03-20
+
+### Changed
+
 - dropped the '\n' for commands
 
 ## [1.0.2] - 2025-03-19
 
 ### Changed
+
 - added extension for registering dependencies so making your own lifetime scope is easier
 
 ## [1.0.1] - 2025-03-19
+
 ### Changed
+
 ~~- fixed package dependency path~~
+
 - nvm, rolled it back
 
 ## [1.0.0] - 2025-03-18